changes for vaultunlocked

+++ src/main/java/nl/trifox/foxprison/economy/Economy.java
@@ -1,10 +1,11 @@
 package nl.trifox.foxprison.economy;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface Economy {
     boolean isAvailable();
-    double getBalance(UUID playerId);
-    boolean withdraw(UUID playerId, double amount);
-    void deposit(UUID playerId, double amount);
+    BigDecimal getBalance(UUID playerId);
+    boolean withdraw(UUID playerId, BigDecimal amount);
+    void deposit(UUID playerId, BigDecimal amount);
 }
